light refactoring, code reduction
