+++ bokeh/sphinxext/bokeh_directive.py
@@ -1,0 +1,30 @@
+'''
+
+'''
+import re
+
+from docutils import nodes
+from docutils.statemachine import ViewList
+
+from sphinx.util.compat import Directive
+from sphinx.util.nodes import nested_parse_with_titles
+
+# taken from Sphinx autodoc
+py_sig_re = re.compile(
+    r'''^ ([\w.]*\.)?            # class name(s)
+          (\w+)  \s*             # thing name
+          (?: \((.*)\)           # optional: arguments
+           (?:\s* -> \s* (.*))?  #           return annotation
+          )? $                   # and nothing more
+          ''', re.VERBOSE)
+
+class BokehDirective(Directive):
+
+    def _parse(self, rst_text, annotation):
+        result = ViewList()
+        for line in rst_text.split("\n"):
+            result.append(line, annotation)
+        node = nodes.paragraph()
+        node.document = self.state.document
+        nested_parse_with_titles(self.state, result, node)
+        return node.children
