+++ bokeh/sphinxext/bokeh_options.py
@@ -31,30 +31,17 @@ the above usage yields the output:
 from __future__ import absolute_import, print_function
 
 import importlib
-import re
 import textwrap
 
-from docutils import nodes
 from docutils.parsers.rst.directives import unchanged
-from docutils.statemachine import ViewList
 
 from sphinx.errors import SphinxError
-from sphinx.util.compat import Directive
-from sphinx.util.nodes import nested_parse_with_titles
 
-from ..core.properties import Options
+from ..util.options import Options
+from .bokeh_directive import BokehDirective, py_sig_re
 from .templates import OPTIONS_DETAIL
 
-# taken from Sphinx autodoc
-py_sig_re = re.compile(
-    r'''^ ([\w.]*\.)?            # class name(s)
-          (\w+)  \s*             # thing name
-          (?: \((.*)\)           # optional: arguments
-           (?:\s* -> \s* (.*))?  #           return annotation
-          )? $                   # and nothing more
-          ''', re.VERBOSE)
-
-class BokehOptionsDirective(Directive):
+class BokehOptionsDirective(BokehDirective):
 
     has_content = True
     required_arguments = 1
@@ -100,13 +87,7 @@ class BokehOptionsDirective(Directive):
 
         rst_text = OPTIONS_DETAIL.render(opts=opts)
 
-        result = ViewList()
-        for line in rst_text.split("\n"):
-            result.append(line, "<bokeh-options>")
-        node = nodes.paragraph()
-        node.document = self.state.document
-        nested_parse_with_titles(self.state, result, node)
-        return node.children
+        return self._parse(rst_text, "<bokeh-options>")
 
 def setup(app):
     app.add_directive_to_domain('py', 'bokeh-options', BokehOptionsDirective)
