+++ bokeh/sphinxext/example_handler.py
